+++ javaScript/projects/weather-app/src/app.js
@@ -1,0 +1,40 @@
+import test from './components/test';
+import './app.css';
+import getData from './containers/api_controller';
+
+// testing, global
+window.getData = getData;
+
+test.print();
+document.body.appendChild(test.render());
+
+// getData('london');
+
+// use WeatherAPI
+// https://www.weatherapi.com/docs/
+
+// things to do:
+// You should be able to search for a specific location
+// toggle displaying the data in Fahrenheit or Celsius.
+// You should change the look of the page based on the data, maybe by changing the color of the background or by adding images that describe the weather
+
+// inputs:
+// 1. city or postal codes
+
+// design:
+// add a ‘loading’ component that displays from the time the form is submitted until the information comes back from the API.
+// 3 day forecast
+// hourly and daily forecast
+
+// layout:
+// <app>
+//    <header> (navigation)
+//    <content>
+//      <heading>
+//      <input> (with C/F toggle button)
+//      <output>
+//        <today> (get current date)
+//        <hourly> (get user's current time)
+//          time | conditions | temp | feels like | precip chance % | precip amount % | cloud cover % | dew point | humidity % | wind speed AND direction
+//        <3-day> (3-day forecast)
+//    <footer>