+++ javaScript/projects/weather-app/src/containers/api_controller.js
@@ -1,0 +1,31 @@
+// use WeatherAPI
+// https://www.weatherapi.com/docs/
+// http://api.weatherapi.com/v1/current.json?key=84ac7310e56448a1896212731230611&q=London
+
+export default async function getData(value) {
+  // *note, value does NOT need to be evaluated before fetch
+  // postal code, number or string
+  // city, uppercase or lowercase;
+
+  try {
+    const response = await fetch(
+      // static value
+      // 'http://api.weatherapi.com/v1/current.json?key=84ac7310e56448a1896212731230611&q=London&days=1&aqi=no&alerts=no',
+      // dynamic value
+      `http://api.weatherapi.com/v1/forecast.json?key=84ac7310e56448a1896212731230611&q=${value}&days=3&aqi=no&alerts=no`,
+      // error test
+      // `http://api.weatherapi.com/v1/forecas.json?key=84ac7310e56448a1896212731230611&q=${value}&days=3&aqi=no&alerts=no`,
+      { mode: 'cors' },
+    );
+
+    const weatherData = await response.json();
+    if (!response.ok) {
+      throw new Error(weatherData.error.message);
+    }
+
+    console.log(weatherData);
+    console.log(weatherData.current);
+  } catch (err) {
+    console.log(err);
+  }
+}